+++ private/pkg/protodescriptor/protodescriptor.go
@@ -18,15 +18,19 @@ import (
 	"errors"
 	"fmt"
 	"strconv"
-	"strings"
 
 	"github.com/bufbuild/buf/private/pkg/normalpath"
+	"github.com/bufbuild/protocompile"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/descriptorpb"
 )
 
-// false until buf CLI can fully and correctly support editions
-var allowEditions = false
+const (
+	// MinSupportedEdition is the earliest edition supported by this repo.
+	MinSupportedEdition = descriptorpb.Edition_EDITION_2023
+	// MaxSupportedEdition is the latest edition supported by this repo.
+	MaxSupportedEdition = descriptorpb.Edition_EDITION_2023
+)
 
 // FileDescriptor is an interface that matches the methods on a *descriptorpb.FileDescriptorProto.
 //
@@ -133,18 +137,17 @@ func ValidateFileDescriptor(fileDescriptor FileDescriptor) error {
 	if err := ValidateProtoPaths("FileDescriptor.Dependency", fileDescriptor.GetDependency()); err != nil {
 		return err
 	}
-	// TODO: Remove this once the CLI is a-okay to work with editions.
-	if fileDescriptor.GetSyntax() == "editions" && !allowEditions {
-		return fmt.Errorf("%s uses edition %s, but editions are not yet supported in buf",
-			fileDescriptor.GetName(), strings.TrimPrefix(fileDescriptor.GetEdition().String(), "EDITION_"))
-	}
-	if fileDescriptor.GetSyntax() == "editions" && fileDescriptor.GetEdition() != descriptorpb.Edition_EDITION_2023 {
-		// Currently, we could only support edition 2023.
-		// TODO: For maximum safety, we probably want to ask protocompile if it supports
-		//       the edition, too. It's a little unlikely that we'd update buf CLI to
-		//       support an edition before protocompile, but just in case.
-		return fmt.Errorf("%s uses unsupported edition %s",
-			fileDescriptor.GetName(), fileDescriptor.GetEdition())
+	if fileDescriptor.GetSyntax() == "editions" {
+		edition := fileDescriptor.GetEdition()
+		// protocompile should support the same editions as buf (or possibly a superset at
+		// some point in the future, like while support for a new edition is being implemented),
+		// but we check with it just in case.
+		if !protocompile.IsEditionSupported(edition) ||
+			edition < MinSupportedEdition ||
+			edition > MaxSupportedEdition {
+			return fmt.Errorf("%s uses unsupported edition %s",
+				fileDescriptor.GetName(), edition)
+		}
 	}
 	return nil
 }
@@ -239,16 +242,4 @@ func FieldDescriptorProtoLabelPrettyString(l descriptorpb.FieldDescriptorProto_L
 	default:
 		return strconv.Itoa(int(l))
 	}
-}
-
-// AllowEditionsForTesting enables support for Protobuf editions, which should only
-// be done for testing until buf fully supports editions correctly (at which point
-// this function will be removed).
-//
-// This function and the internal flag it touches are not thread-safe. So this
-// should be used as early as possible, before any other use of this package. It
-// should typically be called from a TestMain function for a package whose tests
-// need to verify editions-specific behavior.
-func AllowEditionsForTesting() {
-	allowEditions = true
 }